# perf: use batched lru.Flights to reduce cache contentions