--- a/lru_test.go
+++ b/lru_test.go
@@ -8,6 +8,8 @@ import (
 	"sync/atomic"
 	"testing"
 	"time"
+
+	"github.com/redis/rueidis/internal/cmds"
 )
 
 const PTTL = 50
@@ -238,6 +240,228 @@ func TestLRU(t *testing.T) {
 			}
 		})
 	})
+
+	t.Run("Batch Cache Hit & Expire", func(t *testing.T) {
+		lru := setup(t)
+		if v, _ := lru.Flight("0", "GET", TTL, time.Now()); v.typ == 0 {
+			t.Fatalf("did not get the value from the second Flight")
+		} else if v.string != "0" {
+			t.Fatalf("got unexpected value from the second Flight: %v", v)
+		}
+		time.Sleep(PTTL * time.Millisecond)
+		if v, entry := flights(lru, time.Now(), TTL, "GET", "0"); v.typ != 0 || entry != nil {
+			t.Fatalf("got unexpected value from the Flight after pttl: %v %v", v, entry)
+		}
+	})
+
+	t.Run("Batch Cache Should Not Expire By PTTL -2", func(t *testing.T) {
+		lru := setup(t)
+		if v, entry := lru.Flight("1", "GET", TTL, time.Now()); v.typ != 0 || entry != nil {
+			t.Fatalf("got unexpected value from the Flight after pttl: %v %v", v, entry)
+		}
+		m := RedisMessage{typ: '+', string: "1"}
+		lru.Update("1", "GET", m)
+		if v, _ := flights(lru, time.Now(), TTL, "GET", "1"); v.typ == 0 {
+			t.Fatalf("did not get the value from the second Flight")
+		} else if v.string != "1" {
+			t.Fatalf("got unexpected value from the second Flight: %v", v)
+		}
+	})
+
+	t.Run("Batch Cache Miss Suppress", func(t *testing.T) {
+		count := 5000
+		lru := setup(t)
+		wg := sync.WaitGroup{}
+		wg.Add(count)
+		for i := 0; i < count; i++ {
+			go func() {
+				defer wg.Done()
+				if v, _ := flights(lru, time.Now(), TTL, "GET", "1"); v.typ != 0 {
+					t.Errorf("got unexpected value from the first Flight: %v", v)
+				}
+				if v, _ := flights(lru, time.Now(), TTL, "GET", "2"); v.typ != 0 {
+					t.Errorf("got unexpected value from the first Flight: %v", v)
+				}
+			}()
+		}
+		wg.Wait()
+		lru.mu.RLock()
+		store1 := lru.store["1"]
+		store2 := lru.store["2"]
+		lru.mu.RUnlock()
+		if miss := atomic.LoadUint64(&store1.miss); miss != 1 {
+			t.Fatalf("unexpected miss count %v", miss)
+		}
+		if hits := atomic.LoadUint64(&store1.hits); hits != uint64(count-1) {
+			t.Fatalf("unexpected hits count %v", hits)
+		}
+		if miss := atomic.LoadUint64(&store2.miss); miss != 1 {
+			t.Fatalf("unexpected miss count %v", miss)
+		}
+		if hits := atomic.LoadUint64(&store2.hits); hits != uint64(count-1) {
+			t.Fatalf("unexpected hits count %v", hits)
+		}
+	})
+
+	t.Run("Batch Cache Evict", func(t *testing.T) {
+		lru := setup(t)
+		for i := 1; i <= Entries; i++ {
+			flights(lru, time.Now(), TTL, "GET", strconv.Itoa(i))
+			m := RedisMessage{typ: '+', string: strconv.Itoa(i)}
+			m.setExpireAt(time.Now().Add(PTTL * time.Millisecond).UnixMilli())
+			lru.Update(strconv.Itoa(i), "GET", m)
+		}
+		if v, entry := flights(lru, time.Now(), TTL, "GET", "1"); v.typ != 0 {
+			t.Fatalf("got evicted value from the first Flight: %v %v", v, entry)
+		}
+		if v, _ := flights(lru, time.Now(), TTL, "GET", strconv.Itoa(Entries)); v.typ == 0 {
+			t.Fatalf("did not get the latest value from the Flight")
+		} else if v.string != strconv.Itoa(Entries) {
+			t.Fatalf("got unexpected value from the Flight: %v", v)
+		}
+	})
+
+	t.Run("Batch Cache Delete", func(t *testing.T) {
+		lru := setup(t)
+		lru.Delete([]RedisMessage{{string: "0"}})
+		if v, _ := flights(lru, time.Now(), TTL, "GET", "0"); v.typ != 0 {
+			t.Fatalf("got unexpected value from the first Flight: %v", v)
+		}
+	})
+
+	t.Run("Batch Cache Flush", func(t *testing.T) {
+		lru := setup(t)
+		for i := 1; i < Entries; i++ {
+			flights(lru, time.Now(), TTL, "GET", strconv.Itoa(i))
+			m := RedisMessage{typ: '+', string: strconv.Itoa(i)}
+			lru.Update(strconv.Itoa(i), "GET", m)
+		}
+		for i := 1; i < Entries; i++ {
+			if v, _ := flights(lru, time.Now(), TTL, "GET", strconv.Itoa(i)); v.string != strconv.Itoa(i) {
+				t.Fatalf("got unexpected value before flush all: %v", v)
+			}
+		}
+		lru.Delete(nil)
+		for i := 1; i <= Entries; i++ {
+			if v, _ := flights(lru, time.Now(), TTL, "GET", strconv.Itoa(i)); v.typ != 0 {
+				t.Fatalf("got unexpected value after flush all: %v", v)
+			}
+		}
+	})
+
+	t.Run("Batch Cache Close", func(t *testing.T) {
+		lru := setup(t)
+		v, entry := flights(lru, time.Now(), TTL, "GET", "1")
+		if v.typ != 0 || entry != nil {
+			t.Fatalf("got unexpected value from the first Flight: %v %v", v, entry)
+		}
+		v, entry = flights(lru, time.Now(), TTL, "GET", "1")
+		if v.typ != 0 || entry == nil { // entry should not be nil in second call
+			t.Fatalf("got unexpected value from the second Flight: %v %v", v, entry)
+		}
+
+		lru.Close(ErrDoCacheAborted)
+
+		if _, err := entry.Wait(context.Background()); err != ErrDoCacheAborted {
+			t.Fatalf("got unexpected value after Close: %v", err)
+		}
+
+		m := RedisMessage{typ: '+', string: "this Update should have no effect"}
+		m.setExpireAt(time.Now().Add(PTTL * time.Millisecond).UnixMilli())
+		lru.Update("1", "GET", m)
+		for i := 0; i < 2; i++ { // entry should be always nil after the first call if Close
+			if v, entry := flights(lru, time.Now(), TTL, "GET", "1"); v.typ != 0 || entry != nil {
+				t.Fatalf("got unexpected value from the first Flight: %v %v", v, entry)
+			}
+		}
+	})
+
+	t.Run("Batch Cache Cancel", func(t *testing.T) {
+		lru := setup(t)
+		v, entry := flights(lru, time.Now(), TTL, "GET", "1")
+		if v.typ != 0 || entry != nil {
+			t.Fatalf("got unexpected value from the first Flight: %v %v", v, entry)
+		}
+		v, entry = flights(lru, time.Now(), TTL, "GET", "1")
+		if v.typ != 0 || entry == nil { // entry should not be nil in second call
+			t.Fatalf("got unexpected value from the second Flight: %v %v", v, entry)
+		}
+		err := errors.New("any")
+
+		go func() {
+			lru.Cancel("1", "GET", err)
+		}()
+
+		if _, err2 := entry.Wait(context.Background()); err2 != err {
+			t.Fatalf("got unexpected value from the CacheEntry.Wait(): %v %v", err, err2)
+		}
+	})
+
+	t.Run("Batch GetTTL", func(t *testing.T) {
+		lru := setup(t)
+		if v := lru.GetTTL("empty", "cmd"); v != -2 {
+			t.Fatalf("unexpected %v", v)
+		}
+		flights(lru, time.Now(), time.Second, "cmd", "key")
+		m := RedisMessage{typ: 1}
+		m.setExpireAt(time.Now().Add(time.Second).UnixMilli())
+		lru.Update("key", "cmd", m)
+		if v := lru.GetTTL("key", "cmd"); !roughly(v, time.Second) {
+			t.Fatalf("unexpected %v", v)
+		}
+	})
+
+	t.Run("Batch Update Message TTL", func(t *testing.T) {
+		t.Run("client side TTL > server side TTL", func(t *testing.T) {
+			lru := setup(t)
+			flights(lru, time.Now(), time.Second*2, "cmd", "key")
+			m := RedisMessage{typ: 1}
+			m.setExpireAt(time.Now().Add(time.Second).UnixMilli())
+			lru.Update("key", "cmd", m)
+			if v, _ := flights(lru, time.Now(), time.Second*2, "cmd", "key"); v.CacheTTL() != 1 {
+				t.Fatalf("unexpected %v", v.CacheTTL())
+			}
+		})
+		t.Run("client side TTL < server side TTL", func(t *testing.T) {
+			lru := setup(t)
+			flights(lru, time.Now(), time.Second*2, "cmd", "key")
+			m := RedisMessage{typ: 1}
+			m.setExpireAt(time.Now().Add(3 * time.Second).UnixMilli())
+			lru.Update("key", "cmd", m)
+			if v, _ := flights(lru, time.Now(), time.Second*2, "cmd", "key"); v.CacheTTL() != 2 {
+				t.Fatalf("unexpected %v", v.CacheTTL())
+			}
+		})
+		t.Run("no server side TTL -1", func(t *testing.T) {
+			lru := setup(t)
+			flights(lru, time.Now(), time.Second*2, "cmd", "key")
+			m := RedisMessage{typ: 1}
+			lru.Update("key", "cmd", m)
+			if v, _ := flights(lru, time.Now(), time.Second*2, "cmd", "key"); v.CacheTTL() != 2 {
+				t.Fatalf("unexpected %v", v.CacheTTL())
+			}
+		})
+		t.Run("no server side TTL -2", func(t *testing.T) {
+			lru := setup(t)
+			flights(lru, time.Now(), time.Second*2, "cmd", "key")
+			m := RedisMessage{typ: 1}
+			lru.Update("key", "cmd", m)
+			if v, _ := flights(lru, time.Now(), time.Second*2, "cmd", "key"); v.CacheTTL() != 2 {
+				t.Fatalf("unexpected %v", v.CacheTTL())
+			}
+		})
+	})
+}
+
+func flights(lru *lru, now time.Time, ttl time.Duration, args ...string) (RedisMessage, CacheEntry) {
+	results := make([]RedisResult, 1)
+	entries := make(map[int]CacheEntry, 1)
+	lru.Flights(now, commands(ttl, args...), results, entries)
+	return results[0].val, entries[0]
+}
+
+func commands(ttl time.Duration, args ...string) []CacheableTTL {
+	return []CacheableTTL{CT(Cacheable(cmds.NewCompleted(args)), ttl)}
 }
 
 func roughly(ttl, expect time.Duration) bool {