Improve throughput by 13% to 42%

```
▶ benchstat old.txt new.txt
name                                 old time/op    new time/op    delta
/OneNode/128_RueidisMGetCache-10       11.7µs ± 2%    10.2µs ± 0%  -13.50%  (p=0.000 n=10+8)
/OneNode/128_RueidisDoMultiCache-10    7.75µs ± 1%    4.58µs ± 1%  -40.85%  (p=0.000 n=10+9)
/Cluster/128_RueidisMGetCache-10       12.9µs ± 4%    10.9µs ± 1%  -15.36%  (p=0.000 n=10+8)
/Cluster/128_RueidisDoMultiCache-10    9.00µs ± 3%    5.22µs ± 4%  -42.07%  (p=0.000 n=10+10)

name                                 old alloc/op   new alloc/op   delta
/OneNode/128_RueidisMGetCache-10       31.0kB ± 1%    31.1kB ± 1%     ~     (p=0.754 n=10+10)
/OneNode/128_RueidisDoMultiCache-10    11.9kB ± 1%    11.8kB ± 1%   -1.31%  (p=0.000 n=10+10)
/Cluster/128_RueidisMGetCache-10       31.2kB ± 1%    31.0kB ± 1%     ~     (p=0.052 n=10+10)
/Cluster/128_RueidisDoMultiCache-10    11.7kB ± 1%    11.7kB ± 1%     ~     (p=0.224 n=10+10)

name                                 old allocs/op  new allocs/op  delta
/OneNode/128_RueidisMGetCache-10         21.0 ± 0%      21.0 ± 0%     ~     (all equal)
/OneNode/128_RueidisDoMultiCache-10      17.0 ± 0%      17.0 ± 0%     ~     (all equal)
/Cluster/128_RueidisMGetCache-10         20.0 ± 0%      20.0 ± 0%     ~     (all equal)
/Cluster/128_RueidisDoMultiCache-10      16.0 ± 0%      16.0 ± 0%     ~     (all equal)
```

Benchmark source:
```go
package sep

import (
	"context"
	"math/rand"
	"strconv"
	"testing"
	"time"

	"github.com/redis/rueidis"
)

var charset = []byte("abcdefghijklmnopqrstuvwxyz")

func randstr(n int) string {
	b := make([]byte, n)
	for i := range b {
		b[i] = charset[rand.Intn(len(charset))]
	}
	return string(b)
}

func Benchmark(b *testing.B) {
	testfn := func(b *testing.B, n int, addresses []string) {
		ns := strconv.Itoa(n)
		makeclient := func(addresses []string) rueidis.Client {
			client, err := rueidis.NewClient(rueidis.ClientOption{
				InitAddress: addresses,
			})
			if err != nil {
				panic(err)
			}
			return client
		}

		// prepare keys
		keys := make([]string, n)
		cmds := make(rueidis.Commands, n)

		{
			client := makeclient(addresses)
			for i := range keys {
				keys[i] = randstr(10)
				cmds[i] = client.B().Set().Key(keys[i]).Value(randstr(50)).Build()
			}
			client.Do(context.Background(), client.B().Flushall().Build())
			resps := client.DoMulti(context.Background(), cmds...)
			if len(resps) != len(keys) {
				panic("worn")
			}
			client.Close()
		}
		b.Run(ns+" RueidisMGetCache", func(b *testing.B) {
			client := makeclient(addresses)
			defer client.Close()
			b.ReportAllocs()
			b.ResetTimer()
			b.RunParallel(func(pb *testing.PB) {
				for pb.Next() {
					ret, err := rueidis.MGetCache(client, context.Background(), time.Minute, keys)
					if err != nil || len(ret) != len(keys) {
						panic(err)
					}
				}
			})
			b.StopTimer()
		})
		b.Run(ns+" RueidisDoMultiCache", func(b *testing.B) {
			client := makeclient(addresses)
			defer client.Close()
			commands := make([]rueidis.CacheableTTL, len(keys))
			for i := range commands {
				commands[i] = rueidis.CT(client.B().Get().Key(keys[i]).Cache().Pin(), time.Minute)
			}
			b.ReportAllocs()
			b.ResetTimer()
			b.RunParallel(func(pb *testing.PB) {
				for pb.Next() {
					ret := client.DoMultiCache(context.Background(), commands...)
					if len(ret) != len(keys) {
						panic("wrong")
					}
				}
			})
			b.StopTimer()
		})
	}

	b.Run("OneNode", func(b *testing.B) {
		for _, n := range []int{128} {
			testfn(b, n, []string{"127.0.0.1:6379"})
		}
	})

	b.Run("Cluster", func(b *testing.B) {
		for _, n := range []int{128} {
			testfn(b, n, []string{"127.0.0.1:7001", "127.0.0.1:7002", "127.0.0.1:7003"})
		}
	})
}

```